refactor #714 AllDiscussion 패키지 제거 및 분리
