--- android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/compose/discussion/all/AllDiscussionMode.kt
@@ -1,6 +1,0 @@
-package com.team.todoktodok.presentation.compose.discussion.all
-
-enum class AllDiscussionMode {
-    LATEST,
-    SEARCH,
-}
