+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/compose/discussion/model/AllDiscussionMode.kt
@@ -1,0 +1,6 @@
+package com.team.todoktodok.presentation.compose.discussion.model
+
+enum class AllDiscussionMode {
+    LATEST,
+    SEARCH,
+}
