--- android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/compose/preview/AllDiscussionUiStatePreviewParameterProvider.kt
@@ -1,19 +1,0 @@
-package com.team.todoktodok.presentation.compose.preview
-
-import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import com.team.todoktodok.presentation.compose.discussion.all.AllDiscussionMode
-import com.team.todoktodok.presentation.compose.discussion.all.AllDiscussionsUiState
-
-class AllDiscussionUiStatePreviewParameterProvider : PreviewParameterProvider<AllDiscussionsUiState> {
-    override val values: Sequence<AllDiscussionsUiState>
-        get() =
-            sequenceOf(
-                AllDiscussionsUiState(
-                    mode = AllDiscussionMode.LATEST,
-                ),
-                AllDiscussionsUiState(
-                    mode = AllDiscussionMode.SEARCH,
-                    searchDiscussion = SearchDiscussionsUiStatePreviewParameterProvider().values.first(),
-                ),
-            )
-}
